// src/Lineage_old/types.generated.ts
export type Maybe<T> = T | null;
// ...
export type BrowsableEntity = {
  /** The browse paths corresponding to an entity. If no Browse Paths have been generated before, this will be null. */
  browsePaths?: Maybe<Array<BrowsePath>>;
};
// ...
export type BrowsePath = {
  __typename?: 'BrowsePath';
  /** The components of the browse path */
  path: Array<Scalars['String']>;
};

export type BrowsePathEntry = {
  __typename?: 'BrowsePathEntry';
  /**
   * An optional entity associated with this browse entry. This will usually be a container entity.
   * If this entity is not populated, the name must be used.
   */
  entity?: Maybe<Entity>;
  /** The path name of a group of browse results */
  name: Scalars['String'];
};

/** A hierarchical entity path V2 */
export type BrowsePathV2 = {
  __typename?: 'BrowsePathV2';
  /** The components of the browse path */
  path: Array<BrowsePathEntry>;
};
// ...
export type Container = Entity & {
  __typename?: 'Container';
  /** The Roles and the properties to access the container */
  access?: Maybe<Access>;
  /**
   * Experimental API.
   * For fetching extra entities that do not have custom UI code yet
   */
  aspects?: Maybe<Array<RawAspect>>;
  /** The browse path V2 corresponding to an entity. If no Browse Paths V2 have been generated before, this will be null. */
  browsePathV2?: Maybe<BrowsePathV2>;
  /** Fetch an Entity Container by primary key (urn) */
  container?: Maybe<Container>;
  /** The specific instance of the data platform that this entity belongs to */
  dataPlatformInstance?: Maybe<DataPlatformInstance>;
  /** The deprecation status of the container */
  deprecation?: Maybe<Deprecation>;
  /** The Domain associated with the Dataset */
  domain?: Maybe<DomainAssociation>;
  /** Read-write properties that originate in DataHub */
  editableProperties?: Maybe<ContainerEditableProperties>;
  /** Children entities inside of the Container */
  entities?: Maybe<SearchResults>;
  /** Whether or not this entity exists on DataHub */
  exists?: Maybe<Scalars['Boolean']>;
  /** The forms associated with the Dataset */
  forms?: Maybe<Forms>;
  /** The structured glossary terms associated with the dataset */
  glossaryTerms?: Maybe<GlossaryTerms>;
  /** References to internal resources related to the dataset */
  institutionalMemory?: Maybe<InstitutionalMemory>;
  /** The timestamp for the last time this entity was ingested */
  lastIngested?: Maybe<Scalars['Long']>;
  /** Ownership metadata of the dataset */
  ownership?: Maybe<Ownership>;
  /** Recursively get the lineage of containers for this entity */
  parentContainers?: Maybe<ParentContainersResult>;
  /** Standardized platform. */
  platform: DataPlatform;
  /** Privileges given to a user relevant to this entity */
  privileges?: Maybe<EntityPrivileges>;
  /** Read-only properties that originate in the source data platform */
  properties?: Maybe<ContainerProperties>;
  /** Edges extending from this entity */
  relationships?: Maybe<EntityRelationshipsResult>;
  /** Status metadata of the container */
  status?: Maybe<Status>;
  /** Structured properties about this asset */
  structuredProperties?: Maybe<StructuredProperties>;
  /** Sub types of the container, e.g. "Database" etc */
  subTypes?: Maybe<SubTypes>;
  /** Tags used for searching dataset */
  tags?: Maybe<GlobalTags>;
  /** A standard Entity Type */
  type: EntityType;
  /** The primary key of the container */
  urn: Scalars['String'];
};
// ...
export type DataJob = BrowsableEntity & Entity & EntityWithRelationships & {
  __typename?: 'DataJob';
  /**
   * Experimental API.
   * For fetching extra entities that do not have custom UI code yet
   */
  aspects?: Maybe<Array<RawAspect>>;
  /** The browse path V2 corresponding to an entity. If no Browse Paths V2 have been generated before, this will be null. */
  browsePathV2?: Maybe<BrowsePathV2>;
  /** The browse paths corresponding to the data job. If no Browse Paths have been generated before, this will be null. */
  browsePaths?: Maybe<Array<BrowsePath>>;
  /** The parent container in which the entity resides */
  container?: Maybe<Container>;
  /**
   * Deprecated, use relationship IsPartOf instead
   * The associated data flow
   */
  dataFlow?: Maybe<DataFlow>;
  /** The specific instance of the data platform that this entity belongs to */
  dataPlatformInstance?: Maybe<DataPlatformInstance>;
  /** Data Transform Logic associated with the Data Job */
  dataTransformLogic?: Maybe<DataTransformLogic>;
  /** The deprecation status of the Data Flow */
  deprecation?: Maybe<Deprecation>;
  /** The Domain associated with the Data Job */
  domain?: Maybe<DomainAssociation>;
  /** Additional read write properties associated with the Data Job */
  editableProperties?: Maybe<DataJobEditableProperties>;
  /** Whether or not this entity exists on DataHub */
  exists?: Maybe<Scalars['Boolean']>;
  /** The forms associated with the Dataset */
  forms?: Maybe<Forms>;
  /**
   * Deprecated, use the tags field instead
   * The structured tags associated with the DataJob
   * @deprecated Field no longer supported
   */
  globalTags?: Maybe<GlobalTags>;
  /** The structured glossary terms associated with the dashboard */
  glossaryTerms?: Maybe<GlossaryTerms>;
  /** Experimental! The resolved health statuses of the asset */
  health?: Maybe<Array<Health>>;
  /** Incidents associated with the DataJob */
  incidents?: Maybe<EntityIncidentsResult>;
  /**
   * Deprecated, use properties field instead
   * Additional read only information about a Data processing job
   * @deprecated Field no longer supported
   */
  info?: Maybe<DataJobInfo>;
  /** Information about the inputs and outputs of a Data processing job including column-level lineage. */
  inputOutput?: Maybe<DataJobInputOutput>;
  /** References to internal resources related to the dashboard */
  institutionalMemory?: Maybe<InstitutionalMemory>;
  /** Id of the job */
  jobId: Scalars['String'];
  /** The timestamp for the last time this entity was ingested */
  lastIngested?: Maybe<Scalars['Long']>;
  /** Edges extending from this entity grouped by direction in the lineage graph */
  lineage?: Maybe<EntityLineageResult>;
  /** Ownership metadata of the job */
  ownership?: Maybe<Ownership>;
  /** Recursively get the lineage of containers for this entity */
  parentContainers?: Maybe<ParentContainersResult>;
  /** Privileges given to a user relevant to this entity */
  privileges?: Maybe<EntityPrivileges>;
  /** Additional read only properties associated with the Data Job */
  properties?: Maybe<DataJobProperties>;
  /** Granular API for querying edges extending from this entity */
  relationships?: Maybe<EntityRelationshipsResult>;
  /** History of runs of this task */
  runs?: Maybe<DataProcessInstanceResult>;
  /** Status metadata of the DataJob */
  status?: Maybe<Status>;
  /** Structured properties about this asset */
  structuredProperties?: Maybe<StructuredProperties>;
  /** Sub Types that this entity implements */
  subTypes?: Maybe<SubTypes>;
  /** The tags associated with the DataJob */
  tags?: Maybe<GlobalTags>;
  /** A standard Entity Type */
  type: EntityType;
  /** The primary key of the Data Job */
  urn: Scalars['String'];
};
// ...
export type DataJobEditableProperties = {
  __typename?: 'DataJobEditableProperties';
  /** Description of the Data Job */
  description?: Maybe<Scalars['String']>;
};
// ...
export type DataTransformLogic = {
  __typename?: 'DataTransformLogic';
  /** List of transformations applied */
  transforms: Array<DataTransform>;
};
// ...
export type Entity = {
  /** List of relationships between the source Entity and some destination entities with a given types */
  relationships?: Maybe<EntityRelationshipsResult>;
  /** A standard Entity Type */
  type: EntityType;
  /** A primary key of the Metadata Entity */
  urn: Scalars['String'];
};
// ...
export enum EntityType {
  /** A DataHub Access Token */
  AccessToken = 'ACCESS_TOKEN',
  /** A DataHub Assertion */
  Assertion = 'ASSERTION',
  /** A Business Attribute */
  BusinessAttribute = 'BUSINESS_ATTRIBUTE',
  /** The Chart Entity */
  Chart = 'CHART',
  /** A container of Metadata Entities */
  Container = 'CONTAINER',
  /** The CorpGroup Entity */
  CorpGroup = 'CORP_GROUP',
  /** The CorpUser Entity */
  CorpUser = 'CORP_USER',
  /** A Custom Ownership Type */
  CustomOwnershipType = 'CUSTOM_OWNERSHIP_TYPE',
  /** The Dashboard Entity */
  Dashboard = 'DASHBOARD',
  /** A connection to an external source. */
  DatahubConnection = 'DATAHUB_CONNECTION',
  /** A DataHub Policy */
  DatahubPolicy = 'DATAHUB_POLICY',
  /** A DataHub Role */
  DatahubRole = 'DATAHUB_ROLE',
  /** A DataHub View */
  DatahubView = 'DATAHUB_VIEW',
  /** The Dataset Entity */
  Dataset = 'DATASET',
  /** A data contract */
  DataContract = 'DATA_CONTRACT',
  /** The Data Flow (or Data Pipeline) Entity, */
  DataFlow = 'DATA_FLOW',
  /** The Data Job (or Data Task) Entity */
  DataJob = 'DATA_JOB',
  /** The DataPlatform Entity */
  DataPlatform = 'DATA_PLATFORM',
  /** Data Platform Instance Entity */
  DataPlatformInstance = 'DATA_PLATFORM_INSTANCE',
  /** An instance of an individual run of a data job or data flow */
  DataProcessInstance = 'DATA_PROCESS_INSTANCE',
  /** A Data Product */
  DataProduct = 'DATA_PRODUCT',
  /**
   * "
   * A data type registered to DataHub
   */
  DataType = 'DATA_TYPE',
  /** A Domain containing Metadata Entities */
  Domain = 'DOMAIN',
  /**
   * "
   * A type of entity registered to DataHub
   */
  EntityType = 'ENTITY_TYPE',
  /** The ERModelRelationship Entity */
  ErModelRelationship = 'ER_MODEL_RELATIONSHIP',
  /** A DataHub ExecutionRequest */
  ExecutionRequest = 'EXECUTION_REQUEST',
  /**
   * "
   * A form entity on entities
   */
  Form = 'FORM',
  /** The Glossary Node Entity */
  GlossaryNode = 'GLOSSARY_NODE',
  /** The Glossary Term Entity */
  GlossaryTerm = 'GLOSSARY_TERM',
  /** A DataHub incident - SaaS only */
  Incident = 'INCIDENT',
  /** A DataHub Managed Ingestion Source */
  IngestionSource = 'INGESTION_SOURCE',
  /** The ML Feature Entity */
  Mlfeature = 'MLFEATURE',
  /** ML Feature Table Entity */
  MlfeatureTable = 'MLFEATURE_TABLE',
  /** The ML Model Entity */
  Mlmodel = 'MLMODEL',
  /** The MLModelGroup Entity */
  MlmodelGroup = 'MLMODEL_GROUP',
  /** The ML Primary Key Entity */
  MlprimaryKey = 'MLPRIMARY_KEY',
  /** The Notebook Entity */
  Notebook = 'NOTEBOOK',
  /** Another entity type - refer to a provided entity type urn. */
  Other = 'OTHER',
  /** A DataHub Post */
  Post = 'POST',
  /** A dataset query */
  Query = 'QUERY',
  /**
   * "
   * A type of entity that is restricted to the user
   */
  Restricted = 'RESTRICTED',
  /**
   * "
   * A Role from an organisation
   */
  Role = 'ROLE',
  /** A Schema Field */
  SchemaField = 'SCHEMA_FIELD',
  /**
   * "
   * An structured property on entities
   */
  StructuredProperty = 'STRUCTURED_PROPERTY',
  /** The Tag Entity */
  Tag = 'TAG',
  /** A DataHub Test */
  Test = 'TEST',
  /** A set of versioned entities, representing a single source / logical entity over time */
  VersionSet = 'VERSION_SET'
}
// ...
export type EntityWithRelationships = {
  /** Edges extending from this entity grouped by direction in the lineage graph */
  lineage?: Maybe<EntityLineageResult>;
  /** Granular API for querying edges extending from this entity */
  relationships?: Maybe<EntityRelationshipsResult>;
  /** A standard Entity Type */
  type: EntityType;
  /** A primary key associated with the Metadata Entity */
  urn: Scalars['String'];
};
// ...
export type RawAspect = {
  __typename?: 'RawAspect';
  /** The name of the aspect */
  aspectName: Scalars['String'];
  /** JSON string containing the aspect's payload */
  payload?: Maybe<Scalars['String']>;
  /** Details for the frontend on how the raw aspect should be rendered */
  renderSpec?: Maybe<AspectRenderSpec>;
};
// ...
export type SearchResult = {
  __typename?: 'SearchResult';
  /** The resolved DataHub Metadata Entity matching the search query */
  entity: Entity;
  /** Additional properties about the search result. Used for rendering in the UI */
  extraProperties?: Maybe<Array<ExtraProperty>>;
  /** Insights about why the search result was matched */
  insights?: Maybe<Array<SearchInsight>>;
  /** Matched field hint */
  matchedFields: Array<MatchedField>;
};