// src/Lineage_old/entity/shared/types.ts
export type GenericEntityProperties = {
    urn?: string;
    type?: EntityType;
    name?: Maybe<string>;
    properties?: Maybe<{
        name?: Maybe<string>;
        description?: Maybe<string>;
        qualifiedName?: Maybe<string>;
        sourceUrl?: Maybe<string>;
        sourceRef?: Maybe<string>;
        businessAttributeDataType?: Maybe<string>;
        externalUrl?: Maybe<string>;
    }>;
    globalTags?: Maybe<GlobalTags>;
    glossaryTerms?: Maybe<GlossaryTerms>;
    ownership?: Maybe<Ownership>;
    domain?: Maybe<DomainAssociation>;
    dataProduct?: Maybe<EntityRelationshipsResult>;
    platform?: Maybe<DataPlatform>;
    dataPlatformInstance?: Maybe<DataPlatformInstance>;
    customProperties?: Maybe<CustomPropertiesEntry[]>;
    structuredProperties?: Maybe<StructuredProperties>;
    institutionalMemory?: Maybe<InstitutionalMemory>;
    schemaMetadata?: Maybe<SchemaMetadata>;
    externalUrl?: Maybe<string>;
    entityTypeOverride?: Maybe<string>; // to indicate something is a Stream, View instead of Dataset... etc
    /** Dataset specific- TODO, migrate these out */
    editableSchemaMetadata?: Maybe<EditableSchemaMetadata>;
    editableProperties?: Maybe<DatasetEditableProperties>;
    autoRenderAspects?: Maybe<Array<RawAspect>>;
    lineageUrn?: string; // If set, render this urn's lineage instead if not in separate siblings mode
    lineageSiblingIcon?: string; // If set, render this entity in lineage along with the sibling icon and do not separate siblings in the sidebar
    upstream?: Maybe<EntityLineageResult>;
    downstream?: Maybe<EntityLineageResult>;
    subTypes?: Maybe<SubTypes>;
    entityCount?: number;
    container?: Maybe<Container>;
    health?: Maybe<Array<Health>>;
    status?: Maybe<Status>;
    deprecation?: Maybe<Deprecation>;
    siblings?: Maybe<SiblingProperties>;
    siblingsSearch?: Maybe<ScrollResults>;
    parentContainers?: Maybe<ParentContainersResult>;
    parentDomains?: Maybe<ParentDomainsResult>;
    children?: Maybe<EntityRelationshipsResult>;
    parentNodes?: Maybe<ParentNodesResult>;
    isAChildren?: Maybe<EntityRelationshipsResult>;
    siblingPlatforms?: Maybe<DataPlatform[]>;
    lastIngested?: Maybe<number>;
    inputFields?: Maybe<InputFields>;
    fineGrainedLineages?: Maybe<FineGrainedLineage[]>;
    privileges?: Maybe<EntityPrivileges>;
    embed?: Maybe<Embed>;
    exists?: boolean;
    origin?: Maybe<FabricType>;
    documentation?: Maybe<Documentation>;
    browsePathV2?: Maybe<BrowsePathV2>;
    inputOutput?: Maybe<DataJobInputOutput>;
    forms?: Maybe<Forms>;
    parent?: Maybe<GenericEntityProperties>;
    displayProperties?: Maybe<DisplayProperties>;
    notes?: Maybe<EntityRelationshipsResult>;
    versionProperties?: Maybe<VersionProperties>;

    // Data process instance
    lastRunEvent?: Maybe<DataProcessRunEvent>;
};